Refactor: 모든 컨트롤러 Swagger 응답 형식 추가 및 서비스 로직 개선
모든 컨트롤러에 대해 Ok, Created 응답 외에 다른 응답에 대한 Swagger 설정을 추가하고 서비스 로직을 조금 더 가독성 있게 개선하였다.

+++ scs-backend/src/question/question.service.ts
@@ -1,4 +1,5 @@
 import {
+    ForbiddenException,
     Injectable,
     Logger,
     NotFoundException,
@@ -75,24 +76,63 @@ export class QuestionService {
         isolationLevel: IsolationLevel.REPEATABLE_READ,
     })
     async updateQuestionContent(
+        userId: number,
         questionId: number,
         updateQuestionContentDto: UpdateQuestionContentDto,
     ): Promise<Question> {
+        // find user from DB
+        const writer = await this.userRepository.findUserById(userId);
+
         // extract content from DTO
         const { content } = updateQuestionContentDto;
 
-        // find question and update content
-        return this.questionRepository.findAndUpdateQuestionContent(
-            questionId,
-            content,
-        );
+        // find question from DB
+        const question =
+            await this.questionRepository.findQuestionById(questionId);
+
+        // if question does not exist, it is an error
+        if (!question) {
+            throw new NotFoundException(
+                `Question with id ${questionId} has not been found.`,
+            );
+        }
+
+        // if the question has not been written by user, it is an error
+        if (question.writer.id !== writer.id) {
+            throw new ForbiddenException(
+                "User cannot update the content of the question.",
+            );
+        }
+
+        // update and return question
+        question.content = content;
+        return this.questionRepository.save(question);
     }
 
     // [Q-04] Service logic
     @Transactional({
         isolationLevel: IsolationLevel.REPEATABLE_READ,
     })
-    async deleteQuestion(questionId: number): Promise<void> {
+    async deleteQuestion(userId: number, questionId: number): Promise<void> {
+        // find user from DB
+        const writer = await this.userRepository.findUserById(userId);
+
+        // find question from DB
+        const question =
+            await this.questionRepository.findQuestionById(questionId);
+
+        // if the question has not been found, it is an error
+        if (!question) {
+            throw new NotFoundException(
+                `Question with id ${questionId} has not been found.`,
+            );
+        }
+
+        // if the question has not been written by user, it is an error
+        if (question.writer.id !== writer.id) {
+            throw new ForbiddenException("User cannot delete the question.");
+        }
+
         return this.questionRepository.deleteQuestionById(questionId);
     }
